Add method to update velocity for a list of deconfliction types
This CL moves the logic to update velocity of nodes inside of the
deconfliction interface, as an initial step to decoupling collision
avoidance algorithms from the `run_lego_demo` body.

We also change the deconfliction_type parameter into an array of Symbols
that take in a list of deconfliction strategies that should be applied
to the simulation.

Bug: #48
Test: All tests pass
Test Summary:          | Pass  Total     Time
ConstructionBots Tests |   15     15  3m38.5s
     Testing ConstructionBots tests passed