Create constants.jl and define simulation constants
These constants are not be directly tied to RVO, as they pertain to the
simulation rather than type of deconfliction algorithm used. We create a
new `constants.jl` file to fix and globally access these constants as
desired.

Defining them in a separate file increases clarity and ease of
maintenance.

Test: Ran all tests
Bug: #51